VDD measure fixed

+++ samples/tag_battery/src/battery.h
@@ -9,6 +9,6 @@ void battery_init();
 
 void battery_start();
 
-uint16_t battery_get_mv();
+int32_t battery_get_mv();
 
 #endif /*__BATTERY_H__*/
